# stonks_java/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Stonks Simulator API
  version: 1.0.0
  description: COBOL-powered meme stock trading simulator API
  contact:
    name: Pollito
    url: https://springboot.pollito.tech/about/about-the-author
servers:
  - url: http://localhost:8080/api
    description: dev
  - url: https://stonks-java.pollito.tech/api
    description: prod
tags:
  - name: Trades
    description: Endpoints related to trade validation, execution, and paper tape
  - name: Stocks
    description: Endpoints related to stock market data and live prices
  - name: Portfolio
    description: Endpoints related to player portfolio and holdings
  - name: ChaoticEvents
    description: Endpoints related to AI-generated chaotic trading events
  - name: IntensityLevel
    description: Endpoints related to market intensity/volatility level management
paths:
  /trades:
    post:
      tags:
        - Trades
      operationId: executeTrade
      summary: Executes a buy/sell trade
      description: |
        Validates and executes a trade via COBOL PORTFOLIO-MGR,
        persists the updated portfolio state, and returns the result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeExecutionRequest'
      responses:
        '200':
          description: Trade execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeExecutionResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Trades
      operationId: getTradeHistory
      summary: Returns paginated trade history
      description: |
        Returns a paginated list of past trades for the current portfolio.
        Ordered by execution time descending by default.
      x-spring-paginated: true
      responses:
        '200':
          description: Paginated trade history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeHistoryResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /stocks:
    get:
      tags:
        - Stocks
      operationId: getStocks
      summary: Lists all meme stocks with current prices
      description: |
        Returns the full catalog of 10 meme stocks with their latest simulated prices,
        price changes, and change percentages.
      responses:
        '200':
          description: Current market stock prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StocksResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /portfolio:
    get:
      tags:
        - Portfolio
      operationId: getPortfolio
      summary: Returns current portfolio state
      description: |
        Returns the player's current cash balance, stock positions, and unrealized P&L.
        Position valuations are computed using the latest market prices.
      responses:
        '200':
          description: Current portfolio state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /chaotic-events:
    get:
      tags:
        - ChaoticEvents
      operationId: getChaoticEvents
      summary: Lists all active chaotic trading events
      description: |
        Returns a list of all currently active chaotic trading events generated by the AI chaos engine.
      responses:
        '200':
          description: List of active chaotic trading events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaoticEventsResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - ChaoticEvents
      operationId: triggerChaoticEvent
      summary: Triggers a new AI-generated chaotic trading event
      description: |
        Requests the AI chaos engine to generate and activate a new chaotic trading event
        that will affect market dynamics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaoticEventTriggerRequest'
      responses:
        '200':
          description: Chaotic trading event triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaoticEventTriggeredResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /intensity-level:
    get:
      tags:
        - IntensityLevel
      operationId: getIntensityLevel
      summary: Returns the current intensity level
      description: |
        Returns the current intensity level, which controls event frequency and market volatility.
      responses:
        '200':
          description: Current intensity level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntensityLevelResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - IntensityLevel
      operationId: setIntensityLevel
      summary: Sets the intensity level
      description: |
        Sets the intensity level, which controls event frequency and market volatility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntensityLevelSetRequest'
      responses:
        '200':
          description: Intensity level set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntensityLevelResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Error:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          description: Standard error response
          properties:
            title:
              type: string
              description: HTTP Reason Phrase
              example: 'Not Found'
            detail:
              type: string
              description: Description of the problem
              example: "No static resource for request '/'."
          required:
            - title
    TradeAction:
      type: string
      enum: [BUY, SELL]
      description: Trade direction
      example: BUY
    TradeExecutionRequest:
      type: object
      description: Trade to be executed
      properties:
        action:
          $ref: '#/components/schemas/TradeAction'
        symbol:
          type: string
          description: Stock ticker symbol
          example: GMEE
          maxLength: 4
        quantity:
          type: integer
          description: Number of shares
          example: 10
          minimum: 0
      required:
        - action
        - symbol
        - quantity
    TradeExecutionResult:
      type: object
      description: Result returned after executing a trade
      properties:
        status:
          type: string
          enum: [ACCEPTED, REJECTED]
          description: Execution status
          example: ACCEPTED
        errorCode:
          type: string
          description: COBOL-style error code when rejected; null when accepted
          example: S222
          nullable: true
        message:
          type: string
          description: Human-readable execution message
          example: "TRADE EXECUTED - BUY GMEE"
          nullable: true
        newCashBalance:
          type: number
          format: double
          multipleOf: 0.01
          description: Cash balance after the trade
          example: 9550.00
          minimum: 0
        newQuantity:
          type: integer
          description: Updated position quantity after the trade
          example: 10
          minimum: 0
        totalCost:
          type: number
          format: double
          multipleOf: 0.01
          description: Total cost of the trade (quantity * price)
          example: 450.00
          minimum: 0
      required:
        - status
        - newCashBalance
        - newQuantity
        - totalCost
    TradeExecutionResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is the trade execution result
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TradeExecutionResult'
          required:
            - data
    TradeHistoryItem:
      type: object
      description: A single trade history record
      properties:
        id:
          type: integer
          format: int64
          description: Record identifier
          example: 1
        action:
          type: string
          description: Trade direction
          example: BUY
        symbol:
          type: string
          description: Stock ticker symbol
          example: GMEE
          maxLength: 4
        quantity:
          type: integer
          description: Number of shares
          example: 10
        price:
          type: number
          format: double
          multipleOf: 0.01
          description: Price per share at execution
          example: 45.00
          minimum: 0
        totalCost:
          type: number
          format: double
          multipleOf: 0.01
          description: Total cost (quantity * price)
          example: 450.00
          minimum: 0
        cashBalanceAfter:
          type: number
          format: double
          multipleOf: 0.01
          description: Cash balance after the trade
          example: 9550.00
          minimum: 0
        executedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the trade was executed
          example: '2026-01-03T17:11:50.826722328Z'
      required:
        - id
        - action
        - symbol
        - quantity
        - price
        - totalCost
        - cashBalanceAfter
        - executedAt
    TradeHistoryResponseData:
      description: Paginated trade history data
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TradeHistoryItem'
    TradeHistoryResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is the paginated trade history
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TradeHistoryResponseData'
          required:
            - data
    StockPrice:
      type: object
      description: Current price snapshot for a single stock
      properties:
        symbol:
          type: string
          description: Stock ticker symbol
          example: GMEE
          maxLength: 4
        name:
          type: string
          description: Full stock name
          example: GameStonks
        price:
          type: number
          format: double
          multipleOf: 0.01
          description: Current simulated price
          example: 47.85
          minimum: 0
        change:
          type: number
          format: double
          multipleOf: 0.01
          description: Absolute price change since last tick
          example: 2.35
        changePercent:
          type: number
          format: double
          multipleOf: 0.01
          description: Percentage price change since last tick
          example: 5.17
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the price snapshot
          example: '2026-01-03T17:11:50.826722328Z'
      required:
        - symbol
        - name
        - price
        - change
        - changePercent
        - timestamp
    StocksResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is an array of stock prices
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/StockPrice'
          required:
            - data
    Position:
      type: object
      description: A single stock position in the portfolio
      properties:
        symbol:
          type: string
          description: Stock ticker symbol
          example: GMEE
          maxLength: 4
        quantity:
          type: integer
          description: Number of shares held
          example: 10
        currentPrice:
          type: number
          format: double
          multipleOf: 0.01
          description: Latest market price per share
          example: 47.85
          minimum: 0
        marketValue:
          type: number
          format: double
          multipleOf: 0.01
          description: Total market value (quantity * currentPrice)
          example: 478.50
          minimum: 0
        costBasis:
          type: number
          format: double
          multipleOf: 0.01
          description: Weighted average cost per share
          example: 45.00
          minimum: 0
        unrealizedPnl:
          type: number
          format: double
          multipleOf: 0.01
          description: Unrealized profit or loss for this position
          example: 28.50
      required:
        - symbol
        - quantity
        - currentPrice
        - marketValue
        - costBasis
        - unrealizedPnl
    Portfolio:
      type: object
      description: Player's current portfolio state
      properties:
        cashBalance:
          type: number
          format: double
          multipleOf: 0.01
          description: Available cash balance
          example: 9550.00
          minimum: 0
        positions:
          type: array
          description: Current stock holdings
          items:
            $ref: '#/components/schemas/Position'
        unrealizedPnl:
          type: number
          format: double
          multipleOf: 0.01
          description: Total unrealized profit or loss across all positions
          example: 28.50
      required:
        - cashBalance
        - positions
        - unrealizedPnl
    PortfolioResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is the portfolio state
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Portfolio'
          required:
            - data
    Page:
      type: object
      description: Sublist of a list of objects. It allows to gain information about the position of it in the containing entire list
      properties:
        content:
          default: []
          items: {}
          type: array
        pageable:
          $ref: "#/components/schemas/Pageable"
        totalElements:
          default: 0
          description: Total number of items that meet the criteria
          example: 10
          type: integer
        totalPages:
          default: 0
          description: Total pages of items that meet the criteria
          example: 10
          type: integer
      required:
        - content
        - pageable
        - totalElements
        - totalPages
    Pageable:
      type: object
      description: Pagination information
      properties:
        pageNumber:
          description: Current page number (starts from 0)
          example: 0
          minimum: 0
          type: integer
        pageSize:
          description: Number of items retrieved on this page
          example: 10
          minimum: 0
          type: integer
      required:
        - pageNumber
        - pageSize
    ResponseMetadata:
      type: object
      description: Standard response
      properties:
        instance:
          type: string
          description: API endpoint that was called
          example: '/api/something'
        status:
          type: integer
          description: HTTP status code
          example: 200
          minimum: 100
          maximum: 599
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the response was generated
          example: '2026-01-03T17:11:50.826722328Z'
        trace:
          type: string
          description: Unique trace identifier for debugging purposes. Under unexpected situations it can be '00000000000000000000000000000000'
          example: '9482c151-b417-43ff-9dbb-ee12b84e5d99'
      required:
        - instance
        - timestamp
        - trace
        - status
    ChaoticEventSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
      description: Severity level of a chaos event
      example: HIGH
    ChaoticEventType:
      type: string
      enum: [HYPE_WAVE, MEME_STORM, DUMP, WHALE_ALERT, RUG_PULL, PUMP_AND_DUMP, NEWS_FLASH]
      description: Type of chaos event affecting the market
      example: HYPE_WAVE
    ChaoticEvent:
      type: object
      description: A single chaos event affecting market dynamics
      properties:
        eventId:
          type: string
          format: uuid
          description: Unique identifier for the chaos event
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type:
          $ref: '#/components/schemas/ChaoticEventType'
        severity:
          $ref: '#/components/schemas/ChaoticEventSeverity'
        title:
          type: string
          description: Human-readable title of the chaos event
          example: "GMEE Hype Wave Incoming!"
        description:
          type: string
          description: Detailed description of the chaos event
          example: "A wave of retail hype is driving GMEE prices up. Expect increased volatility."
        targetSymbol:
          type: string
          description: Stock ticker symbol affected by the event, if any
          example: GMEE
          maxLength: 4
          nullable: true
        affectedSymbols:
          type: array
          items:
            type: string
          description: Stock ticker symbols affected by the event
          example: ["GMEE", "FOMO"]
          nullable: true
        priceEffect:
          type: number
          format: double
          description: Estimated price effect multiplier (e.g. 1.15 for +15%)
          example: 1.15
        startedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the event started
        expiresAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the event expires
      required:
        - eventId
        - type
        - severity
        - title
        - description
        - startedAt
        - expiresAt
    ChaoticEventsResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is a list of chaos events
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ChaoticEvent'
          required:
            - data
    ChaoticEventTriggerRequest:
      type: object
      description: Request to trigger a new chaos event
      properties:
        type:
          $ref: '#/components/schemas/ChaoticEventType'
        severity:
          $ref: '#/components/schemas/ChaoticEventSeverity'
        targetSymbol:
          type: string
          description: Stock ticker symbol to target, if any
          example: GMEE
          maxLength: 4
          nullable: true
      required:
        - type
        - severity
    IntensityLevel:
      type: string
      enum: [PAPER_HANDS, MODERATE, HIGH_VOLATILITY, EXTREME, MAXIMUM_OVERDRIVE]
      description: Chaos level controlling event frequency and volatility
      example: PAPER_HANDS
    IntensityLevelSetRequest:
      type: object
      description: Request body for setting the chaos level
      properties:
        level:
          $ref: '#/components/schemas/IntensityLevel'
      required:
        - level
    IntensityLevelResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is the chaos level
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/IntensityLevel'
          required:
            - data
    ChaoticEventTriggeredResponse:
      x-implements:
        - dev.pollito.stonks_java.util.metadata.ResponseMetadata
      description: Standard response where the data property is the triggered chaos event
      allOf:
        - $ref: '#/components/schemas/ResponseMetadata'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ChaoticEvent'
          required:
            - data